refactor: cleaning code
