+++ src/Components/SingleWinner.scss
@@ -1,0 +1,11 @@
+.winners{
+    &__table{
+      &--header{
+      background-color: #4D96FF;
+      }
+      &--cell{
+        padding:10px 20px;
+      }
+  
+    }
+  }
