+++ src/Winners.jsx
@@ -1,9 +1,12 @@
 import './App.scss';
-import { LittleCarGenerater } from './Helpers/LittleCarGenerate';
+import { SingleWinner } from './Components/SingleWinner';
+
 export const Winners = ({winners, cars})=>{
 
     return <div className="winners">
+        
         <h2 className="garage__title">Winners({winners?.length})</h2>
+
         <table>
             <thead  className='winners__table--header'>
                 <tr>
@@ -16,13 +19,11 @@ export const Winners = ({winners, cars})=>{
             </thead>
             <tbody>
                 {winners?.map((winner, i)=>{
-                    return <tr key={i}>
-                        <td className='winners__table--cell'>{i+1}</td>
-                        <td className='winners__table--cell'>{LittleCarGenerater(cars?.find(car=>car?.id===winner?.id)?.color)}</td>
-                        <td className='winners__table--cell'>{cars?.find(car=>car?.id===winner?.id)?.name}</td>
-                        <td className='winners__table--cell'>{winner?.wins}</td>
-                        <td className='winners__table--cell'>{winner?.time}</td>
-                    </tr>
+                    return <SingleWinner 
+                            winner={winner} 
+                            i={i} 
+                            cars={cars}
+                            />
                 })}
             </tbody>
         </table>
